% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{condition}
\alias{condition}
\title{Title for condition}
\format{
The conditions of bulk sample state from TCGA-CHOL,including tumor and normal.
}
\usage{
condition
}
\description{
The conditions of bulk sample state or cell annotation of single-cell transcriptomes.
}
\keyword{datasets}
