% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{exp_matrix}
\alias{exp_matrix}
\title{Title for exp_matrix}
\format{
The expression matrix of 45 samples and 20531 genes of TCGA-CHOL normal and tumor.
}
\usage{
exp_matrix
}
\description{
The expression matrix of the samples or single cells for which enrichment scoring needs to be performed.
}
\keyword{datasets}
