% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdGSE.r
\name{gdGSE}
\alias{gdGSE}
\title{gdGSE: a novel pathway enrichment evaluation algorithm based on discretization of gene expression values}
\usage{
gdGSE(exp_matrix, condition, Signature, data_type)
}
\arguments{
\item{exp_matrix}{The expression matrix of the samples or single cells for which enrichment scoring needs to be performed}

\item{condition}{The conditions of bulk "Tumor" or "Normal" or cell annotation of single-cell transcriptomes.}

\item{Signature}{A list of signatures,for example:the Gene Ontology Biological Process(GOBP) gene sets, cell type-specific signatures or a gene list correlated with interested features.}

\item{data_type}{"Bulk" or "SingleCell".}
}
\value{
Description of the return value.
}
\description{
gdGSE: a novel pathway enrichment evaluation algorithm based on discretization of gene expression values
}
