% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{SC_condition}
\alias{SC_condition}
\title{Title for SC_condition}
\format{
The expression matrix of 1000 cells and 5000 genes from GSE149614.
}
\usage{
SC_condition
}
\description{
The cell annotation of the single cells for from GSE149614.
}
\keyword{datasets}
