% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{SC_exp_matrix}
\alias{SC_exp_matrix}
\title{Title for SC_exp_matrix}
\format{
The expression matrix of 1000 cells and 5000 genes from GSE149614.
}
\usage{
SC_exp_matrix
}
\description{
The expression matrix of the single cells for which enrichment scoring needs to be performed.
}
\keyword{datasets}
