% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{SC_Signature}
\alias{SC_Signature}
\title{Title for SC_Signature}
\format{
The cell marker gene sets of the single cells for from GSE149614.
}
\usage{
SC_Signature
}
\description{
The cell marker gene sets of the single cells for from GSE149614.
}
\keyword{datasets}
