% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Signature}
\alias{Signature}
\title{Title for Signature}
\format{
100 Gene Ontology Biological Process(GOBP) gene sets.
}
\usage{
Signature
}
\description{
A list of signatures,for example:the Gene Ontology Biological Process(GOBP) gene sets, cell type-specific signatures or a gene list correlated with interested features.
}
\keyword{datasets}
